fix malloc'ed buffer sizes in CcspCrProcessDbusRequest()
The val array contains pointers to data of type name_spaceType_t, so
the buffers allocated for each pointer should be of size
sizeof(name_spaceType_t). Multiplying by param_size seems to be a
historical copy and paste error from the code above it used to
allocate the val array.